# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /
    description: API root
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Vapp operational overview
  - name: products
    description: Product management
paths:
  /health:
    get:
      operationId: rootHealthCheck
      tags: [health]
      summary: Root health check
      description: Returns safe service health metadata
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /v1/health:
    get:
      operationId: versionedHealthCheck
      tags: [health]
      summary: Versioned health check
      description: Returns safe service health metadata
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /ready:
    get:
      operationId: rootReadinessCheck
      tags: [health]
      summary: Root readiness check
      description: Returns readiness based on MongoDB availability
      responses:
        "200":
          description: Ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessStatus"
        "503":
          description: Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessStatus"
  /v1/ready:
    get:
      operationId: versionedReadinessCheck
      tags: [health]
      summary: Versioned readiness check
      description: Returns readiness based on MongoDB availability
      responses:
        "200":
          description: Ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessStatus"
        "503":
          description: Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessStatus"
  /v1/dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get Vapp dashboard counts
      responses:
        "200":
          description: Current domain counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /v1/products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List and filter products
      parameters:
        - in: query
          name: search
          schema:
            type: string
            maxLength: 200
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/ProductStatus"
      responses:
        "200":
          description: Products ordered by most recently updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/products/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          pattern: "^product_[0-7][0-9abcdefghjkmnpqrstvwxyz]{25}$"
    get:
      operationId: getProduct
      tags: [products]
      summary: Get one product
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
components:
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ProductStatus:
      type: string
      enum: [idea, validation, active, paused, retired]
    Product:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - slug
        - status
        - productType
        - domains
        - commercialModel
        - oneOffPurchaseAvailable
        - subscriptionAvailable
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          pattern: "^product_[0-7][0-9abcdefghjkmnpqrstvwxyz]{25}$"
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/ProductStatus"
        productType:
          type: string
        domains:
          type: array
          items:
            type: string
        commercialModel:
          type: string
        oneOffPurchaseAvailable:
          type: boolean
        subscriptionAvailable:
          type: boolean
        currency:
          type: string
        internalNotes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      additionalProperties: false
      required: [name, slug, status, productType, commercialModel]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          minLength: 2
          maxLength: 100
          pattern: "^[a-z0-9][a-z0-9-]*$"
        description:
          type: string
          maxLength: 10000
        status:
          $ref: "#/components/schemas/ProductStatus"
        productType:
          type: string
          minLength: 1
          maxLength: 100
        domains:
          type: array
          maxItems: 50
          items:
            type: string
            minLength: 1
            maxLength: 253
        commercialModel:
          type: string
          minLength: 1
          maxLength: 100
        oneOffPurchaseAvailable:
          type: boolean
        subscriptionAvailable:
          type: boolean
        currency:
          type: string
          pattern: "^[A-Za-z]{3}$"
        internalNotes:
          type: string
          maxLength: 20000
    ProductUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          minLength: 2
          maxLength: 100
          pattern: "^[a-z0-9][a-z0-9-]*$"
        description:
          type: string
          maxLength: 10000
        status:
          $ref: "#/components/schemas/ProductStatus"
        productType:
          type: string
          minLength: 1
          maxLength: 100
        domains:
          type: array
          maxItems: 50
          items:
            type: string
            minLength: 1
            maxLength: 253
        commercialModel:
          type: string
          minLength: 1
          maxLength: 100
        oneOffPurchaseAvailable:
          type: boolean
        subscriptionAvailable:
          type: boolean
        currency:
          type: string
          pattern: "^[A-Za-z]{3}$"
        internalNotes:
          type: string
          maxLength: 20000
    DashboardSummary:
      type: object
      additionalProperties: false
      required:
        - totalProducts
        - activeProducts
        - draftOrInactiveProducts
        - totalOrganisations
        - totalPeople
        - totalOpportunities
      properties:
        totalProducts:
          type: integer
          minimum: 0
        activeProducts:
          type: integer
          minimum: 0
        draftOrInactiveProducts:
          type: integer
          minimum: 0
        totalOrganisations:
          type: integer
          minimum: 0
        totalPeople:
          type: integer
          minimum: 0
        totalOpportunities:
          type: integer
          minimum: 0
    HealthStatus:
      type: object
      properties:
        status:
          type: string
          enum: [ok]
        serviceName:
          type: string
        environment:
          type: string
          enum: [dev, prod, test, local]
        runtimeMode:
          type: string
          enum: [development, test, production]
        version:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - status
        - serviceName
        - environment
        - runtimeMode
        - version
        - timestamp
    ReadinessStatus:
      type: object
      properties:
        status:
          type: string
          enum: [ready, unavailable]
        serviceName:
          type: string
        environment:
          type: string
          enum: [dev, prod, test, local]
        dependencies:
          type: object
          properties:
            mongodb:
              type: string
              enum: [available, unavailable]
          required:
            - mongodb
        timestamp:
          type: string
          format: date-time
      required:
        - status
        - serviceName
        - environment
        - dependencies
        - timestamp