# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  /** @minimum 0 */
  totalProducts: number;
  /** @minimum 0 */
  activeProducts: number;
  /** @minimum 0 */
  draftOrInactiveProducts: number;
  /** @minimum 0 */
  totalOrganisations: number;
  /** @minimum 0 */
  totalPeople: number;
  /** @minimum 0 */
  totalOpportunities: number;
}