# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatus } from './productStatus';

export interface Product {
  /** @pattern ^product_[0-7][0-9abcdefghjkmnpqrstvwxyz]{25}$ */
  id: string;
  name: string;
  slug: string;
  description?: string;
  status: ProductStatus;
  productType: string;
  domains: string[];
  commercialModel: string;
  oneOffPurchaseAvailable: boolean;
  subscriptionAvailable: boolean;
  currency?: string;
  internalNotes?: string;
  createdAt: Date;
  updatedAt: Date;
}