# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatus } from './productStatus';

export interface ProductInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  /**
     * @minLength 2
     * @maxLength 100
     * @pattern ^[a-z0-9][a-z0-9-]*$
     */
  slug: string;
  /** @maxLength 10000 */
  description?: string;
  status: ProductStatus;
  /**
     * @minLength 1
     * @maxLength 100
     */
  productType: string;
  /**
     * @maxItems 50
     * @items.minLength 1
     * @items.maxLength 253
     */
  domains?: string[];
  /**
     * @minLength 1
     * @maxLength 100
     */
  commercialModel: string;
  oneOffPurchaseAvailable?: boolean;
  subscriptionAvailable?: boolean;
  /** @pattern ^[A-Za-z]{3}$ */
  currency?: string;
  /** @maxLength 20000 */
  internalNotes?: string;
}