# lib/api-zod/src/generated/types/productStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ProductStatus = typeof ProductStatus[keyof typeof ProductStatus];


export const ProductStatus = {
  idea: 'idea',
  validation: 'validation',
  active: 'active',
  paused: 'paused',
  retired: 'retired',
} as const;